Added A-Star sequence diagram and A-Start ppt
The sequence diagram shows the basic interaction between various parts of the main driver, A*, and the maps/data-structures. The power point is for presenting the basic ideas of how we are using path planning techniques to drive the robot. Included in the power point are links to familiarize members with the algorithms and basic building blocks of the problem.